+++ resolver/index.ts
