+++ scripts/tradehistory
