--- src/state.ts
@@ -1,7 +1,0 @@
-import { Orderbook } from "./types";
-
-export type State = { orderbook: Orderbook };
-
-const state: State = { orderbook: {} };
-
-export default state;
