+++ test/end-to-end.ts
@@ -1,37 +1,12 @@
 import assert from "assert";
 import fetch from "node-fetch";
-import express from "express";
-import dotenv from "dotenv";
-import path from "path";
-import start from "../src/start";
-import configure from "../src/configure";
-import fs from "fs";
-import { getHash } from "../src/auth";
-import { OrderBookResponse } from "../src/types";
-
+import { OrderBookResponse } from "../orderbook";
+import { getHash } from "../server/auth";
+import { startServer } from "./util";
+/** */
 describe("end-to-end", () => {
-  before(async () => {
-    // ...
-    dotenv.config({
-      path: fs.existsSync(path.join(process.cwd(), ".env.local"))
-        ? path.join(process.cwd(), ".env.local") // Not in source control
-        : path.join(process.cwd(), ".env"),
-    });
-    const app = express();
-    try {
-      const server = await start(app);
-      await configure(app);
-      const { address, family, port } = server.address() as any;
-      return `Express Listening on [${[address, family, port]}]`;
-    } catch (error) {
-      console.error(error);
-      process.exit(1);
-    }
-  });
-
-  after(() => {
-    process.exit();
-  });
+  before(startServer);
+  after(() => process.exit());
 
   const port = process.env.PORT || "5000";
 
@@ -56,7 +31,7 @@ describe("end-to-end", () => {
     assert.strictEqual(typeof requestid, "string");
   });
 
-  it("orderbook", async () => {
+  it("gets orderbook", async () => {
     const {
       asks,
       bids,
@@ -73,7 +48,7 @@ describe("end-to-end", () => {
     assert.strictEqual(0, bids.length);
   });
 
-  it("tradehistory", async () => {
+  it("gets tradehistory", async () => {
     const x: OrderBookResponse[] = await fetch(
       `http://localhost:${port}/api/BTCZAR/tradehistory`,
       {
