+++ test/index.test.ts
@@ -1,3 +1,4 @@
 /// <reference path="../@types/express/index.d.ts" />
-import "./unit";
+import "./orderbook.unit";
+import "./sort-by-price.unit";
 import "./end-to-end";
