+++ tsconfig.json
@@ -16,5 +16,5 @@
     "noEmit": true,
     "typeRoots": ["./@types", "node_modules/@types", "../node_modules/@types"]
   },
-  "include": ["src"]
+  "include": ["resolver", "server", "store", "util"]
 }
